+++ lib/widgets/viewer/overlay/slideshow_buttons.dart
@@ -29,6 +29,7 @@ class _SlideshowButtonsState extends State<SlideshowButtons> {
   static const List<SlideshowAction> _actions = [
     SlideshowAction.resume,
     SlideshowAction.showInCollection,
+    SlideshowAction.settings,
   ];
   static const double _padding = ViewerButtonRowContent.padding;